+++ azalea-entity/src/enchantments.rs
@@ -1,8 +1,0 @@
-pub fn _get_enchant_level(
-    _enchantment: azalea_registry::Enchantment,
-    _player_inventory: &azalea_inventory::Menu,
-) -> u32 {
-    // TODO
-
-    0
-}