+++ azalea-inventory/src/slot.rs
@@ -1,7 +1,7 @@
 use std::{
     any::Any,
     borrow::Cow,
-    fmt,
+    fmt::{self, Debug},
     io::{self, Cursor, Write},
 };
 
@@ -79,11 +79,10 @@ impl ItemStack {
         }
     }
 
-    /// Get the `kind` of the item in this slot, or
-    /// [`azalea_registry::Item::Air`]
-    pub fn kind(&self) -> azalea_registry::Item {
+    /// Get the `kind` of the item in this slot, or [`Item::Air`]
+    pub fn kind(&self) -> Item {
         match self {
-            ItemStack::Empty => azalea_registry::Item::Air,
+            ItemStack::Empty => Item::Air,
             ItemStack::Present(i) => i.kind,
         }
     }
@@ -152,7 +151,7 @@ impl Serialize for ItemStack {
 #[derive(Debug, Clone, PartialEq, Serialize)]
 pub struct ItemStackData {
     #[serde(rename = "id")]
-    pub kind: azalea_registry::Item,
+    pub kind: Item,
     /// The amount of the item in this slot.
     ///
     /// The count can be zero or negative, but this is rare.
@@ -184,7 +183,7 @@ impl ItemStackData {
 
     /// Check if the count of the item is <= 0 or if the item is air.
     pub fn is_empty(&self) -> bool {
-        self.count <= 0 || self.kind == azalea_registry::Item::Air
+        self.count <= 0 || self.kind == Item::Air
     }
 
     /// Whether this item is the same as another item, ignoring the count.
@@ -222,7 +221,7 @@ impl AzaleaRead for ItemStack {
         if count <= 0 {
             Ok(ItemStack::Empty)
         } else {
-            let kind = azalea_registry::Item::azalea_read(buf)?;
+            let kind = Item::azalea_read(buf)?;
             let component_patch = DataComponentPatch::azalea_read(buf)?;
             Ok(ItemStack::Present(ItemStackData {
                 count,
@@ -449,7 +448,7 @@ impl Clone for DataComponentPatch {
         DataComponentPatch { components }
     }
 }
-impl fmt::Debug for DataComponentPatch {
+impl Debug for DataComponentPatch {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_set().entries(self.components.keys()).finish()
     }