+++ codegen/lib/code/data_components.py
@@ -180,6 +180,7 @@ def update_default_variants(version_id: str):
 use std::collections::HashMap;
 
 use azalea_chat::translatable_component::TranslatableComponent;
+use azalea_core::attribute_modifier_operation::AttributeModifierOperation;
 use azalea_registry::{Attribute, Block, EntityKind, HolderSet, Item, MobEffect, SoundEvent};
 use simdnbt::owned::NbtCompound;